backend/requirements.txt
alembic==1.16.5
blinker==1.9.0
click==8.1.8
et_xmlfile==2.0.0
Flask==2.3.3
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.0.2
openpyxl==3.1.5
pandas==2.3.3
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.44
tomli==2.3.0
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==2.3.7
xlsxwriter==3.2.9
zipp==3.23.0
